tools/fru-dump: bugfix: fix an argument to please 64-bit builds
fru-dump includes local headers, where recent commit changed fru_alloc
to correctly receive size_t.  fru-dump has it's own local implementation,
which was still receiving an "int" argument. Thus build failed in 64-bit
(and I wonder why my compiler doesn't even warn in 32 bits.

But reported and fix suggested by Michel Vincent, in private email.

+++ tools/fru-dump.c
@@ -18,7 +18,7 @@
 
 #include "../kernel/fru-parse.c" /* Aaaargh!!!!! horrible hack... */
 
-void *fru_alloc(int size)
+void *fru_alloc(size_t size)
 {
 	return malloc(size);
 }
